refactor(solid-query): separate queryOptions from createQuery
