+++ packages/solid-query/src/createInfiniteQuery.ts
@@ -2,7 +2,6 @@ import { InfiniteQueryObserver } from '@tanstack/query-core'
 import { createMemo } from 'solid-js'
 import { createBaseQuery } from './createBaseQuery'
 import type {
-  DataTag,
   DefaultError,
   InfiniteData,
   QueryKey,
@@ -13,111 +12,12 @@ import type {
   CreateInfiniteQueryOptions,
   CreateInfiniteQueryResult,
   DefinedCreateInfiniteQueryResult,
-  FunctionedParams,
-  SolidInfiniteQueryOptions,
 } from './types'
 import type { Accessor } from 'solid-js'
-
-export type UndefinedInitialDataInfiniteOptions<
-  TQueryFnData,
-  TError = DefaultError,
-  TData = InfiniteData<TQueryFnData>,
-  TQueryKey extends QueryKey = QueryKey,
-  TPageParam = unknown,
-> = FunctionedParams<
-  SolidInfiniteQueryOptions<
-    TQueryFnData,
-    TError,
-    TData,
-    TQueryFnData,
-    TQueryKey,
-    TPageParam
-  > & {
-    initialData?: undefined
-  }
->
-type NonUndefinedGuard<T> = T extends undefined ? never : T
-export type DefinedInitialDataInfiniteOptions<
-  TQueryFnData,
-  TError = DefaultError,
-  // should we handle page param correctly
-  TData = InfiniteData<TQueryFnData>,
-  TQueryKey extends QueryKey = QueryKey,
-  TPageParam = unknown,
-> = FunctionedParams<
-  SolidInfiniteQueryOptions<
-    TQueryFnData,
-    TError,
-    TData,
-    TQueryFnData,
-    TQueryKey,
-    TPageParam
-  > & {
-    initialData:
-      | NonUndefinedGuard<InfiniteData<TQueryFnData, TPageParam>>
-      | (() => NonUndefinedGuard<InfiniteData<TQueryFnData, TPageParam>>)
-  }
->
-function infiniteQueryOptions<
-  TQueryFnData,
-  TError = DefaultError,
-  TData = InfiniteData<TQueryFnData>,
-  TQueryKey extends QueryKey = QueryKey,
-  TPageParam = unknown,
->(
-  options: ReturnType<
-    DefinedInitialDataInfiniteOptions<
-      TQueryFnData,
-      TError,
-      TData,
-      TQueryKey,
-      TPageParam
-    >
-  >,
-): ReturnType<
-  DefinedInitialDataInfiniteOptions<
-    TQueryFnData,
-    TError,
-    TData,
-    TQueryKey,
-    TPageParam
-  >
-> & {
-  queryKey: DataTag<TQueryKey, InfiniteData<TQueryFnData>>
-}
-function infiniteQueryOptions<
-  TQueryFnData,
-  TError = DefaultError,
-  TData = InfiniteData<TQueryFnData>,
-  TQueryKey extends QueryKey = QueryKey,
-  TPageParam = unknown,
->(
-  options: ReturnType<
-    UndefinedInitialDataInfiniteOptions<
-      TQueryFnData,
-      TError,
-      TData,
-      TQueryKey,
-      TPageParam
-    >
-  >,
-): ReturnType<
-  UndefinedInitialDataInfiniteOptions<
-    TQueryFnData,
-    TError,
-    TData,
-    TQueryKey,
-    TPageParam
-  >
-> & {
-  queryKey: DataTag<TQueryKey, InfiniteData<TQueryFnData>>
-}
-
-function infiniteQueryOptions(options: unknown) {
-  return options
-}
-
-export { infiniteQueryOptions }
+import type {
+  DefinedInitialDataInfiniteOptions,
+  UndefinedInitialDataInfiniteOptions,
+} from './infiniteQueryOptions'
 
 export function createInfiniteQuery<
   TQueryFnData,
